Refactor task handling into a new singleton TaskList class
