+++ src/main/java/paddington/task/Deadline.java
@@ -13,4 +13,9 @@ public class Deadline extends Task {
     public String toString() {
         return "[D]" + super.toString() + " (by: " + by + ")";
     }
+
+    @Override
+    public String saveTask() {
+        return "D" + super.toString() + SEPARATOR + by;
+    }
 }
