+++ src/main/java/paddington/task/Event.java
@@ -14,4 +14,9 @@ public class Event extends Task {
     public String toString() {
         return "[E]" + super.toString() + " (from: " + from + " to: " + to + ")";
     }
+
+    @Override
+    public String saveTask() {
+        return "E" + super.toString() + SEPARATOR + from + " " + to;
+    }
 }
