+++ src/main/java/paddington/task/Todo.java
@@ -10,4 +10,9 @@ public class Todo extends Task {
     public String toString() {
         return "[T]" + super.toString();
     }
+
+    @Override
+    public String saveTask() {
+        return "T" + super.toString();
+    }
 }
